+++ src/redux/features/selectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { RootState } from '@/redux/store';
-import { Product } from '@/interfaces/product';
+import { Product } from '@/types/product';
 
 export const selectAllProducts = createSelector(
   (state: RootState) => state.api.queries["getProducts(undefined)"]?.data,
