+++ src/types/IPageProps.ts
@@ -1,0 +1,5 @@
+export interface IPageProps {
+  params: {
+    lang?: string;
+  };
+}
