+++ src/types/index.ts
@@ -1,5 +1,7 @@
 export * from './HeaderType';
+export * from './IPageProps';
 export * from './IBlogsType';
 export * from './IRecommendationProductProps';
 export * from './IVisibleSubcategory';
 export * from './ISubcategoriesApi';
+export * from './product';
